refactor: Expo 동시 요청 제한으로 인한 chunk 수 변경
