+++ batch/src/main/kotlin/dnd11th/blooming/batch/notification/NotificationScheduler.kt
@@ -5,6 +5,7 @@ import org.springframework.batch.core.Job
 import org.springframework.batch.core.JobParameters
 import org.springframework.batch.core.JobParametersBuilder
 import org.springframework.batch.core.launch.JobLauncher
+import org.springframework.scheduling.annotation.Scheduled
 import org.springframework.stereotype.Component
 import org.springframework.util.StopWatch
 import java.util.concurrent.TimeUnit
@@ -14,6 +15,7 @@ class NotificationScheduler(
     private val jobLauncher: JobLauncher,
     private val notificationJob: Job,
 ) {
+    @Scheduled(cron = "0 0 5-23 * * *")
     fun run() {
         val jobParameters: JobParameters =
             JobParametersBuilder()
